+++ src/index.ts
@@ -1,7 +1,9 @@
+import makeDictionary from './Compile'
 import ConsumeLanguage from './Consumer'
 import withLanguageContext from './Provider'
 
 export {
   ConsumeLanguage,
   withLanguageContext,
+  makeDictionary,
 }