+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.0.2
+
+Introduce helper -function `compile(Array<file?, json?, context?>)` for compile suited dictionary
+ - Tests and basic functionality for helper function
+
 ### 0.0.1
 
 Initial version release